# Create NoPublish sbt plugin