Regarding the PR that disabled javadoc publishing which was merged in order to get publishling to work (see https://github.com/apache/incubator-pekko-grpc/pull/78) I noticed that Pekko core implements a basic `NoPublish` sbt plugin which not only sets `publish / skip := true` but also disables publishing (see https://github.com/apache/incubator-pekko/blob/main/project/Publish.scala#L41C1-L49).

In order to avoid potential publishing issues (and also just out of design, if we do not publish an sbt project there is no point in generating docs) I have just ported this solution to pekko-grpc. There is one nuance just for pekk-grpc, which is `publishLocal / skip := false` for `interopt-tests` which was set in https://github.com/apache/incubator-pekko-grpc/pull/64. In order to verify there wasn't a regression I just used `show` on the various keys so I can confirm it shouldn't break anything, i.e.

## Before this PR
```
[IJ]show publish / skip
[info] maven-plugin / publish / skip
[info] 	false
[info] plugin-tester-java / publish / skip
[info] 	true
[info] benchmarks / publish / skip
[info] 	true
[info] sbt-plugin / publish / skip
[info] 	false
[info] interop-tests / publish / skip
[info] 	true
[info] runtime / publish / skip
[info] 	false
[info] codegen / publish / skip
[info] 	false
[info] plugin-tester-scala / publish / skip
[info] 	true
[info] docs / publish / skip
[info] 	true
[info] scalapb-protoc-plugin / publish / skip
[info] 	false
[info] publish / skip
[info] 	true
[success] Total time: 0 s, completed May 28, 2023 12:17:36 PM
[IJ]
```

```
[IJ]show publishLocal / skip
[info] maven-plugin / publishLocal / skip
[info] 	false
[info] plugin-tester-java / publishLocal / skip
[info] 	true
[info] benchmarks / publishLocal / skip
[info] 	true
[info] sbt-plugin / publishLocal / skip
[info] 	false
[info] interop-tests / publishLocal / skip
[info] 	false
[info] runtime / publishLocal / skip
[info] 	false
[info] codegen / publishLocal / skip
[info] 	false
[info] plugin-tester-scala / publishLocal / skip
[info] 	true
[info] docs / publishLocal / skip
[info] 	true
[info] scalapb-protoc-plugin / publishLocal / skip
[info] 	false
[info] publishLocal / skip
[info] 	true
[success] Total time: 0 s, completed May 28, 2023 12:17:51 PM
[IJ]
```

## With this PR
```
[IJ]show publish / skip
[info] maven-plugin / publish / skip
[info] 	false
[info] plugin-tester-java / publish / skip
[info] 	true
[info] benchmarks / publish / skip
[info] 	true
[info] sbt-plugin / publish / skip
[info] 	false
[info] interop-tests / publish / skip
[info] 	true
[info] runtime / publish / skip
[info] 	false
[info] codegen / publish / skip
[info] 	false
[info] plugin-tester-scala / publish / skip
[info] 	true
[info] docs / publish / skip
[info] 	true
[info] scalapb-protoc-plugin / publish / skip
[info] 	false
[info] publish / skip
[info] 	true
[success] Total time: 0 s, completed May 28, 2023 12:14:45 PM
```

```
[IJ]show publishLocal / skip
[info] maven-plugin / publishLocal / skip
[info] 	false
[info] plugin-tester-java / publishLocal / skip
[info] 	true
[info] benchmarks / publishLocal / skip
[info] 	true
[info] sbt-plugin / publishLocal / skip
[info] 	false
[info] interop-tests / publishLocal / skip
[info] 	false
[info] runtime / publishLocal / skip
[info] 	false
[info] codegen / publishLocal / skip
[info] 	false
[info] plugin-tester-scala / publishLocal / skip
[info] 	true
[info] docs / publishLocal / skip
[info] 	true
[info] scalapb-protoc-plugin / publishLocal / skip
[info] 	false
[info] publishLocal / skip
[info] 	true
[success] Total time: 0 s, completed May 28, 2023 12:15:12 PM
[IJ]
```