@mdedetrich 2023-05-29
> lgtm, but it would be better we have a pekko-build-plugin .

Definitely, and I want to both abstract most of this stuff as well as make it consistent between our pekko projects (currently the build structure for our pekko projects is all over the place). 

Something akin to https://github.com/typelevel/sbt-typelevel but for Pekko would be fantastic.